requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
scipy>=1.10.0

# Training utilities
tqdm>=4.65.0
tensorboard>=2.14.0
scikit-learn>=1.3.0

# Dataset (pinned <4.0: NYUv2 HF dataset uses a loading script)
datasets>=2.14.0,<4.0
h5py>=3.9.0

# Export
onnx>=1.14.0
onnxscript>=0.6.0

# Visualization
matplotlib>=3.7.0

# ── Teacher inference (HPC only, not needed for student training) ──────
# depth-anything-v2          # DA2-Large via torch.hub or HuggingFace
# ultralytics                # YOLOv8 for YOLO-seeded SAM2 prompting
# segment-anything-2         # SAM2-Large for mask refinement
# transformers               # Alternative DA2 loading via HuggingFace
